+++ DTCD-Visualization-BarChart/src/Plugin.js
@@ -12,8 +12,11 @@ import {
 export class Plugin extends PanelPlugin {
   #title;
   #targetName;
-  #dataSource;
+  #dataSourceName;
   #storageSystem;
+  #guid;
+  #eventSystem;
+  #dataSourceSystemGUID;
 
   static getRegistrationMeta() {
     return pluginMeta;
@@ -27,7 +30,10 @@ export class Plugin extends PanelPlugin {
     const dataSourceSystem = new DataSourceSystemAdapter();
 
     eventSystem.registerPluginInstance(this);
+    this.#guid = guid;
+    this.#eventSystem = eventSystem;
     this.#storageSystem = new StorageSystemAdapter();
+    this.#dataSourceSystemGUID = this.getGUID(this.getSystem('DataSourceSystem'));
 
     const { default: VueJS } = this.getDependence('Vue');
 
@@ -39,7 +45,7 @@ export class Plugin extends PanelPlugin {
     this.vueComponent = view.$children[0];
     this.#title = '';
     this.#targetName = '';
-    this.#dataSource = '';
+    this.#dataSourceName = '';
   }
 
   loadData(data) {
@@ -49,8 +55,8 @@ export class Plugin extends PanelPlugin {
 
   processDataSourceEvent(eventData) {
     const { dataSource, status } = eventData;
-    this.#dataSource = dataSource;
-    const data = this.#storageSystem.session.getRecord(this.#dataSource);
+    this.#dataSourceName = dataSource;
+    const data = this.#storageSystem.session.getRecord(this.#dataSourceName);
     this.loadData(data);
   }
 
@@ -68,10 +74,29 @@ export class Plugin extends PanelPlugin {
     }
 
     if (typeof dataSource !== 'undefined') {
-      this.#dataSource = dataSource;
-      const DS = this.getSystem('DataSourceSystem').getDataSource(this.#dataSource);
+      if (this.#dataSourceName) {
+        this.#eventSystem.unsubscribe(
+          this.#dataSourceSystemGUID,
+          'DataSourceStatusUpdate',
+          this.#guid,
+          'processDataSourceEvent',
+          { dataSource: this.#dataSourceName, status: 'success' },
+        );
+      }
+
+      this.#dataSourceName = dataSource;
+
+      this.#eventSystem.subscribe(
+        this.#dataSourceSystemGUID,
+        'DataSourceStatusUpdate',
+        this.#guid,
+        'processDataSourceEvent',
+        { dataSource, status: 'success' }
+      );
+
+      const DS = this.getSystem('DataSourceSystem').getDataSource(this.#dataSourceName);
       if (DS.status === 'success') {
-        const data = this.#storageSystem.session.getRecord(this.#dataSource);
+        const data = this.#storageSystem.session.getRecord(this.#dataSourceName);
         this.loadData(data);
       }
     }
@@ -80,7 +105,7 @@ export class Plugin extends PanelPlugin {
   getPluginConfig() {
     const config = {};
     if (this.#title) config.title = this.#title;
-    if (this.#dataSource) config.dataSource = this.#dataSource;
+    if (this.#dataSourceName) config.dataSource = this.#dataSourceName;
     if (this.#targetName) config.targetName = this.#targetName;
     return config;
   }